requirements.txt
asgiref==3.11.0
black==26.1.0
click==8.3.1
Django==6.0.1
mypy_extensions==1.1.0
packaging==26.0
pathspec==1.0.4
platformdirs==4.5.1
pytokens==0.4.0
sqlparse==0.5.5
structlog==24.1.0
django-request-id==1.0.0
django-prometheus==2.3.1
prometheus_client==0.24.1
